pyproject.toml
[tool.poetry]
name = "readagent"
version = "0.1.0"
description = ""
authors = ["Edward Tsoi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gradio = "^4.29.0"
openai = "^1.16.2"
google-generativeai = "^0.5.0"
pypdfium2 = "^4.29.0"
azure-ai-documentintelligence = "^1.0.0b3"
gradio-pdf = "^0.0.7"
boto3 = "^1.34.84"
botocore = "^1.34.84"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
